moved ICardVisitor and IItemVisitor into visitor package
