+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/TransactionVerify.java
@@ -10,7 +10,7 @@ package com.internetEnemies.combatCritters.Logic;
  * is valid, and performs the necessary operations if it is.
  */
 
-import com.internetEnemies.combatCritters.data.Database;
+import com.internetEnemies.combatCritters.Logic.visitor.IItemVisitor;
 import com.internetEnemies.combatCritters.data.ICardInventory;
 import com.internetEnemies.combatCritters.data.ICurrencyInventory;
 import com.internetEnemies.combatCritters.data.IPackInventory;
@@ -19,7 +19,7 @@ import com.internetEnemies.combatCritters.objects.Currency;
 import com.internetEnemies.combatCritters.objects.ItemCard;
 import com.internetEnemies.combatCritters.objects.Pack;
 
-public class TransactionVerify implements IItemVisitor{
+public class TransactionVerify implements IItemVisitor {
     private final ICardInventory cardInventory;
     private final IPackInventory packInventory;
     private final ICurrencyInventory bank;
