+++ app/src/main/java/com/internetEnemies/combatCritters/objects/Card.java
@@ -9,7 +9,7 @@
 
 package com.internetEnemies.combatCritters.objects;
 
-import com.internetEnemies.combatCritters.Logic.ICardVisitor;
+import com.internetEnemies.combatCritters.Logic.visitor.ICardVisitor;
 
 import java.io.Serializable;
 import java.util.Objects;
