+++ app/src/main/java/com/internetEnemies/combatCritters/objects/IItem.java
@@ -1,6 +1,6 @@
 package com.internetEnemies.combatCritters.objects;
 
-import com.internetEnemies.combatCritters.Logic.IItemVisitor;
+import com.internetEnemies.combatCritters.Logic.visitor.IItemVisitor;
 
 /**
  * IVisitableItem.java
